pyproject.toml
[tool.poetry]
name = "money-flow"
version = "0.1.0"
description = ""
authors = ["marnix le noble <github.com/marnixlenoble>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
flask = "^3.0.0"
bunq-sdk = "^1.14.18"
python-dotenv = "^1.0.0"
requests = "^2.31.0"
firebase-admin = "^6.3.0"


[tool.poetry.group.dev.dependencies]
black = "^23.12.1"
pytest = "^7.4.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
